% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerCalendR.R
\name{summerCampCalendR}
\alias{summerCampCalendR}
\title{Create Camp calendar}
\description{
Create Camp calendar

Create Camp calendar
}
\details{
Helps in the creation of a camp calendar that can be shared or printed.
It estimates the total cost using the inputs. It does not account for
regional holidays.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-summerCampCalendR-new}{\code{summerCampCalendR$new()}}
\item \href{#method-summerCampCalendR-AddCampCost}{\code{summerCampCalendR$AddCampCost()}}
\item \href{#method-summerCampCalendR-ViewCamps}{\code{summerCampCalendR$ViewCamps()}}
\item \href{#method-summerCampCalendR-CreateCalendar}{\code{summerCampCalendR$CreateCalendar()}}
\item \href{#method-summerCampCalendR-AddSummerCampDates}{\code{summerCampCalendR$AddSummerCampDates()}}
\item \href{#method-summerCampCalendR-AddSchool}{\code{summerCampCalendR$AddSchool()}}
\item \href{#method-summerCampCalendR-CalculatePrice}{\code{summerCampCalendR$CalculatePrice()}}
\item \href{#method-summerCampCalendR-GetDates}{\code{summerCampCalendR$GetDates()}}
\item \href{#method-summerCampCalendR-GetEvents}{\code{summerCampCalendR$GetEvents()}}
\item \href{#method-summerCampCalendR-GetFirstWorkDayOfWeek}{\code{summerCampCalendR$GetFirstWorkDayOfWeek()}}
\item \href{#method-summerCampCalendR-Plot}{\code{summerCampCalendR$Plot()}}
\item \href{#method-summerCampCalendR-clone}{\code{summerCampCalendR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-new"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-new}{}}}
\subsection{Method \code{new()}}{
Initialize the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-AddCampCost"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-AddCampCost}{}}}
\subsection{Method \code{AddCampCost()}}{
Add a camp and a cost
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$AddCampCost(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{a camp is put in via `Name` = `cost` arguments that are
internally converted to a list and then a `data.table`. Note: Camps can
be overriden in this manner.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-ViewCamps"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-ViewCamps}{}}}
\subsection{Method \code{ViewCamps()}}{
View the camps you have input. You cannot adjust them here.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$ViewCamps()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-CreateCalendar"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-CreateCalendar}{}}}
\subsection{Method \code{CreateCalendar()}}{
Create the calendar internally so you may add events
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$CreateCalendar(start_date, end_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date, end_date}}{must be convertible to `date` class with
`start_date` < `end_date`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-AddSummerCampDates"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-AddSummerCampDates}{}}}
\subsection{Method \code{AddSummerCampDates()}}{
Add the date of the summer camp
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$AddSummerCampDates(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{input as `FirstDate` = `Camp`. The entire week will be filled.
No half weeks allowed at the moment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-AddSchool"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-AddSchool}{}}}
\subsection{Method \code{AddSchool()}}{
add the school dates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$AddSchool(start_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{the first day of school}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-CalculatePrice"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-CalculatePrice}{}}}
\subsection{Method \code{CalculatePrice()}}{
Calculate the approximate price including taxes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$CalculatePrice(tax = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tax}}{the amout of the tax rate. Defaults to `1`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-GetDates"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-GetDates}{}}}
\subsection{Method \code{GetDates()}}{
get the dates provided in the creation of the calendar
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$GetDates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-GetEvents"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-GetEvents}{}}}
\subsection{Method \code{GetEvents()}}{
Get a `data.table` with the dates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$GetEvents()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-GetFirstWorkDayOfWeek"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-GetFirstWorkDayOfWeek}{}}}
\subsection{Method \code{GetFirstWorkDayOfWeek()}}{
get the first work day of the week
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$GetFirstWorkDayOfWeek()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-Plot"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-Plot}{}}}
\subsection{Method \code{Plot()}}{
generate a `gg` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$Plot(palette.name = "Paired", addprice = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{palette.name}}{the name of the RColorBrewer palette. Defaults to 'Paired'}

\item{\code{addprice}}{boolean}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summerCampCalendR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-summerCampCalendR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{summerCampCalendR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
